Small refactor.
